% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spamming.R
\docType{methods}
\name{spamming,lngCMatrix,missing-method}
\alias{spamming,lngCMatrix,missing-method}
\title{Calculate Hamming Distance for All Rows of Matrix}
\usage{
\S4method{spamming}{lngCMatrix,missing}(x, y)
}
\arguments{
\item{x}{A sparse matrix of class \code{lgCMatrix} or \code{ngCMatrix}.}
}
\value{
A matrix of dimension \code{nrow(x), nrow(x)} where each entry
is the hamming distance between row i and row j. See details for current
status.
}
\description{
Calculate Hamming Distance for All Rows of Matrix
}
\details{
The matrix returned by this method is currently an upper
triangle matrix with no entries along the diagonal and all zeros below
the diagonal. This is a WIP and eventually an object of class \code{dist}
will be returned.

The current workaround is to run this code: \code{dst <- d + t(d)}. where
\code{d} is the matrix currently returned by this method.
}
