% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spamming.R
\docType{methods}
\name{spamming,lngCMatrix,lngCMatrix-method}
\alias{spamming,lngCMatrix,lngCMatrix-method}
\title{Calculate Hamming Distance for All Rows of Matrix}
\usage{
\S4method{spamming}{lngCMatrix,lngCMatrix}(x, y)
}
\arguments{
\item{x}{A sparse matrix of class \code{lgCMatrix} or \code{ngCMatrix}.}

\item{y}{A sparse matrix of class \code{lgCMatrix} or \code{ngCMatrix}.}
}
\value{
A matrix of dimension \code{nrow(x), nrow(y)} where each entry
is the hamming distance between row i of \code{x} and row j of \code{y}.
}
\description{
Calculate Hamming Distance for All Rows of Matrix
}
\details{
This method will stop immediately if \code{!identical(ncol(x), ncol(y))}.
}
