% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spamming.R
\docType{methods}
\name{spamming,lngCMatrix,missing-method}
\alias{spamming,lngCMatrix,missing-method}
\title{Calculate Hamming Distance for All Rows of Matrix}
\usage{
\S4method{spamming}{lngCMatrix,missing}(x, y)
}
\arguments{
\item{x}{A sparse matrix of class \code{lgCMatrix} or \code{ngCMatrix}.}
}
\value{
A \code{\link[stats]{dist}} object of size \code{nrow(x), nrow(x)}
where each entry is the hamming distance between row i and row j. See details
for current status.
}
\description{
Calculate Hamming Distance for All Rows of Matrix
}
